% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eee2eye.R
\name{eee2eye}
\alias{eee2eye}
\title{Calculate the E:I for a Steady-State Lake using 18O-H2O}
\usage{
eee2eye(df, dL, dI, dP, temp, humid, evap, k)
}
\arguments{
\item{df}{dataframe with your values}

\item{dL}{steady-state lake isotope value (‰)}

\item{dI}{input/source water isotope value (‰)}

\item{dP}{precipitation isotope value (‰)}

\item{temp}{temperature at surface of lake (C)}

\item{humid}{relative humidity (decimal)}

\item{evap}{annual evaporation rate of area (m/yr)}

\item{k}{influence of seasonality (0.5: highly seasonal; 1: no seasonality)}
}
\description{
This function allows you to calculate the evaporation to inflow (E:I) ratio of a lake based on a few assumptions and its water isotope value.
}
\examples{
eee2eye(ei_input, 'd18O_H2O_permille', 'dI_permille', 'dP_permille', 'temp_C', 'h_dec', 'evap_m.yr', 'k')
}
\keyword{E:I,}
\keyword{balance}
\keyword{isotope}
\keyword{isotopes,}
\keyword{mass}
\keyword{water}
