% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eee2eye_WRT.R
\name{eee2eye_WRT}
\alias{eee2eye_WRT}
\title{Calculate the Water Residence Time from E:I for a Steady-State Lake}
\usage{
eee2eye_WRT(df, E.I, e, SA, V)
}
\arguments{
\item{df}{dataframe with your values}

\item{E.I}{E:I ratio}

\item{e}{annual evaporation depth (m/yr)}

\item{SA}{lake surface area (m2)}

\item{V}{lake volume (m3)}
}
\description{
This function allows you to calculate the water residence time of a lake based on an evaporation to inflow (E:I) ratio.
}
\examples{
ei_input <- data.frame(E.I = c(0.2042, 0.3138, 0.1838),
                       e_myr = c(0.3965, 0.3965, 0.3965),
                       SA_m2 = c(315900, 300825, 589950),
                       V_m3 = c(2466000, 3004064, 5712829))
eee2eye_WRT(ei_input, 'E.I', 'e_myr', 'SA_m2', 'V_m3')
}
\keyword{E:I,}
\keyword{balance}
\keyword{isotope}
\keyword{isotopes,}
\keyword{mass}
\keyword{residence}
\keyword{time,}
\keyword{water}
